refactor: move input_dataframe_generator to tests, coicop/matching/statshelpers/variables to policy
- input_dataframe_generator: moved to tests/ (used only by tests)
- coicop, matching, statshelpers, variables: moved to policy/
- update all imports accordingly
- coicop: paths -> configuration.paths

Made-with: Cursor
